fix: [#59905] Fix pagination
The table-cookie tries to remember the pagination and drop the user off
there. The plugin does not work well cross-table: Cookies for one table
are wrongfully being applied on another. Removing the plugin resolves
the problem.

The plugin was only installed to handle proper table-updating  after
bulk-approval. This is not an issue any more so the plugin can safely be
removed.