fix: [#59971] Import API data from latest "to date" until today
The current import code only imports from latest "to date" plus 1 day.
This causes problems when there are no data in the Tomra API for a given
day.
The next run of the import will use the *same* "from date" and "to date"
as the preceding run, yielding no data, and therefore a new "to date" is
not recorded. This continues at the next run, etc.

We have observed multiple days where no data is available in the API,
due to nothing being scanned, etc.

This causes the daily imports to "stall" in the manner described above.

This commit changes the behavior in two ways:
- the default behavior is now to import date from the "latest to date"
until *today*, which hopefully should mean that the import will "catch
up" once data becomes available again, following one or more days
without data.
- the "from date" and "to date" can be overridden via command line
options, in case a manual operation is necessary anyway.