feat: [#60187] Replace `Product.approved`, etc. with `Product.state` FSM
This drops the following fields from `Product`:
- `approved`
- `closed` (never used in production anyway)
- `approval_date` (we now use the date from first history entry where
the product was approved instead)
- `creation_date` (we now use the date from the first history entry
where the product was created instead)

This adds the following annotations to the default `Product.objects`
queryset (via the new `ProductManager`):
- `awaiting_approval` - True iff the product is awaiting approval
- `approved` - True iff the product is approved; replaces the former
field of the same name
- `rejected` - True iff the product is rejected
- `deleted` - True iff the product is deleted; this means that we no
longer delete products from the database, but only mark them as deleted
(soft delete.)

A new data migration ensures correspondence between `Product.state` and
the corresponding history entries in `HistoricalProduct`:
- all products now have an initial history entry holding their creation
date and user
- products which have been approved also have a history entry holding
the date and user of the approval

The database constraints on `Product` have been updated:
- `barcode` is only unique when a product is not deleted - in other
words, once a product is deleted, another product can be created with
the same barcode.

`Product.history` now has an explicit FK to `Product` instances, as
`related_name` is defined.

This adds `django-fsm` as a new dependency.