test: [#60453] Update tests
Tests now cover:
- deposit payout items without a product,
- deposit payout items without a source (company branch or kiosk)

Tests now assert that:
- deposit payout items belonging to the same source (company branch or
kiosk), and resulting in the same output lines, are grouped together
regardless of whether the underlying objects are linked to a product or
not.
- some items must *not* be marked with a file ID, even though they match
the initial filtering on from/to date, etc. Specifically, items without
a source must not be marked with a file ID, while items without a
product, or manually created items, must be marked with a file ID.