fix: [#62927] Restore show/hide button logic based on selected rows count
The `check.bs.table`, etc. events in `bootstrap-table` are not emitted
when CSP is set up.

The code now relies on our own custom table event `row-select-event`,
and checks whether *all* rows are selected by looking at the
`total-rows` data property.

When selecting a subset of the rows displayed, the ID list is now passed
as a single comma-separated string, rather than a POST array. So the
backend code is also updated to reflect this.