feat: [#64548] Additional `QRBagSearchView` columns
This changes the set of columns on "Pantposer" (`QRBagSearchView`) to
match the mockup provided by the customer.

This updates the `QRBagSearchView` unittests to use a status name
matching one of those used in the actual production database
(`butik_oprettet`) in order to provide test coverage of the code
generating the content for the date columns.

This also tweaks the CSS for tables to allow table headers (column
titles) to word wrap, in order to provide a bit saner layout for very
wide tables with many columns.